fixed extra topping bugs
in java .toLowerCase() creates a new object, so comparing to "y" always comes back false because they are not the same object. Adding .equals solves that.

the Pizza addToppings method did not store whether the topping had extra.

finally, the Topping loop in addPizza was manually setting all toppings hasExtra to false. Now it references the isToppingDoubled array to pass the boolean to useTopping.